pyproject.toml
[tool.ruff]
line-length = 100
target-version = "py310"
extend-exclude = ["testmcpy/research"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by formatter
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "E722",  # bare except - used intentionally in some places
    "B904",  # raise from within except - style preference
    "B028",  # warnings.warn without stacklevel
    "UP007", # Use X | Y instead of Optional - breaks Typer on Python 3.10
    "UP045", # Use X | None instead of Optional - breaks Typer on Python 3.10
]

[tool.ruff.lint.per-file-ignores]
"testmcpy/cli.py" = ["F401"]  # Imports used for availability checks
"tests/test_url_protection.py" = ["F841"]  # Intentional variable assignment for documentation

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["unit_tests", "integration_tests"]
python_files = ["test_*.py"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "e2e: End-to-end UI tests (requires playwright)",
]

[tool.coverage.run]
source = ["testmcpy"]
omit = [
    "tests/*",
    "testmcpy/ui/*",
    "testmcpy/research/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.black]
line-length = 100

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["testmcpy*"]

[tool.setuptools.package-data]
testmcpy = [
    "ui/**/*",
    "server/**/*",
]

[project]
name = "testmcpy"
version = "0.2.17"
description = "A comprehensive testing framework for validating LLM tool calling capabilities with MCP services"
authors = [{name = "Amin Ghadersohi"}]
license = "Apache-2.0"
readme = "README.md"
requires-python = ">=3.10,<3.13"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "typer>=0.9.0,<1.0.0",
    "rich>=13.0.0,<15.0.0",
    "pyyaml>=6.0,<7.0",
    "requests>=2.28.0,<3.0.0",
    "aiohttp>=3.8.0,<4.0.0",
    "ollama>=0.1.0",
    "anthropic>=0.39.0,<1.0.0",
    "fastmcp>=2.0.0,<3.0.0",
    "httpx>=0.27.0,<1.0.0",
    "python-dotenv>=1.0.0,<2.0.0",
    "click>=8.0.0,<9.0.0",
    "shellingham>=1.3.0,<2.0.0",
    "textual>=0.47.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "pre-commit>=3.0.0",
    "build>=1.0.0",
    "twine>=5.0.0",
    "types-pyyaml>=6.0.0",
    "types-requests>=2.28.0",
    "textual-dev>=1.0.0",
]
server = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "websockets>=14.0,<15.0",
]
sdk = [
    "claude-agent-sdk>=0.1.0",
]
tui = [
    "textual>=0.85.0",
]
e2e = [
    "playwright>=1.40.0",
    "pytest-playwright>=0.4.0",
]
all = [
    "fastapi>=0.104.0,<1.0.0",
    "uvicorn[standard]>=0.24.0,<1.0.0",
    "websockets>=14.0,<15.0",
    "claude-agent-sdk>=0.1.0",
    "textual>=0.85.0",
]

[project.scripts]
testmcpy = "testmcpy.cli:app"

[project.urls]
Homepage = "https://github.com/preset-io/testmcpy"
Repository = "https://github.com/preset-io/testmcpy"
Issues = "https://github.com/preset-io/testmcpy/issues"